requirements.txt
# PowerShell에서
New-Item requirements.txt
```

**requirements.txt 내용:**
```
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0